move abstractions around
